Cleaned up boxing warnings.
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@8054 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/tool/user/menus/FileMenu.java
@@ -1644,7 +1644,7 @@ public class FileMenu {
             return JOptionPane.CLOSED_OPTION;
         if(options == null) {
             if(selectedValue instanceof Integer)
-                return ((Integer)selectedValue); // using autoboxing
+                return (((Integer)selectedValue).intValue()); // using autoboxing
             return JOptionPane.CLOSED_OPTION;
         }
         for(int counter = 0, maxCounter = options.length;
